backend/requirements-lock.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.110.1
gpt4all==2.2.1.post1
greenlet==3.0.3
gTTS==2.5.1
h11==0.14.0
idna==3.6
pip==24.0
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
python-dotenv==1.0.1
python-multipart==0.0.9
pyttsx3==2.90
requests==2.31.0
setuptools==59.6.0
sniffio==1.3.1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.41.2
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
wheel==0.37.1